Moved schemeswitching-data-serializer.h from pke/include/scheme/ckksrns to pke/unittest as it is a collection of special utilities (#916)
* Moved schemeswitching-data-serializer.h from pke/include/scheme/ckksrns to pke/unittest as it is a collection of special utilities

* Addressed code review comments

---------
