+++ src/pke/CMakeLists.txt
@@ -63,8 +63,7 @@ if( BUILD_UNITTESTS )
 	endif()
 	add_executable (pke_tests ${PKE_TEST_SRC_FILES} ${UNITTESTMAIN})
 	set_property(TARGET pke_tests PROPERTY RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/unittest)
-    target_include_directories(pke_tests PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/unittest")
-    target_include_directories(pke_tests PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/unittest/utils")
+	target_include_directories(pke_tests PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/unittest/utils")
 	target_link_libraries ( pke_tests ${PKELIBS} ${ADDITIONAL_LIBS})
 	if (NOT ${WITH_OPENMP} )
 		target_link_libraries ( pke_tests PRIVATE Threads::Threads)
@@ -80,7 +79,14 @@ if ( BUILD_EXAMPLES)
 	file (GLOB PKE_EXAMPLES_SRC_FILES CONFIGURE_DEPENDS examples/*.cpp)
 	foreach (app ${PKE_EXAMPLES_SRC_FILES})
 		get_filename_component ( exe ${app} NAME_WE )
-		add_executable ( ${exe} ${app} )
+		if(${exe} STREQUAL "scheme-switching-serial")
+			# add schemeswitching-data-serializer.cpp to link line and
+			# include the unittest/utils directory where schemeswitching-data-serializer.h is
+			add_executable ( ${exe} ${app} "${CMAKE_CURRENT_SOURCE_DIR}/unittest/utils/schemeswitching-data-serializer.cpp")
+			target_include_directories( ${exe} PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/unittest/utils")
+		else()
+			add_executable ( ${exe} ${app} )
+		endif()
 		set_property(TARGET ${exe} PROPERTY RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin/examples/pke)
 		set( PKEAPPS ${PKEAPPS} ${exe} )
 		target_link_libraries ( ${exe} ${PKELIBS} ${ADDITIONAL_LIBS})
